Get test coverage in AutoRestReportTest.swift
